Replace old 'pay later' code by skipping payment for netxx purchases
Needs tests! This was a bit of a spike.

It is easier to accomplish this by overriding #payment_required? on
Order. Also, remind wholesaler to send payment on the order#show page.

Still need to allow site to enter an address to which to mail payment,
in order to show on the order page.